#16 refactor: Member Entity 및 Buider 패턴 리팩토링
