+++ src/main/java/KNUHR/Server/member/domain/Member.java
@@ -1,5 +1,6 @@
-package KNUHR.Server.user.domain;
+package KNUHR.Server.member.domain;
 
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,7 +10,7 @@ import javax.persistence.*;
 import java.util.UUID;
 
 @Entity
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Table(name = "MEMBER")
 public class Member {
