+++ src/main/java/KNUHR/Server/member/dto/LoginRequest.java
@@ -1,0 +1,18 @@
+package KNUHR.Server.member.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginRequest {
+    private String email;
+    private String password;
+
+    @Builder
+    public LoginRequest(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+}
