+++ src/main/java/KNUHR/Server/member/dto/RegisterRequest.java
@@ -1,12 +1,14 @@
-package KNUHR.Server.user.dto;
+package KNUHR.Server.member.dto;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
 
 @Getter
+@NoArgsConstructor
 public class RegisterRequest {
 
     @NotNull
