+++ src/main/java/KNUHR/Server/member/dto/SendEmailRequest.java
@@ -1,12 +1,14 @@
-package KNUHR.Server.user.dto;
+package KNUHR.Server.member.dto;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
 
 @Getter
+@NoArgsConstructor
 public class SendEmailRequest {
     @NotNull
     @Email
@@ -15,9 +17,5 @@ public class SendEmailRequest {
     @Builder
     public SendEmailRequest(@NotNull @Email String email) {
         this.email = email;
-    }
-
-    public SendEmailRequest() {
-
     }
 }
