+++ src/main/java/KNUHR/Server/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package KNUHR.Server.user.repository;
+package KNUHR.Server.member.repository;
 
-import KNUHR.Server.user.domain.Member;
+import KNUHR.Server.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
