--- src/main/java/KNUHR/Server/user/dto/LoginRequest.java
@@ -1,11 +1,0 @@
-package KNUHR.Server.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequest {
-    private String email;
-    private String password;
-}
